+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- hnefatafl-client: Make users and games scrollable.
 - Make all the features dependencies.
 
 [unreleased]: https://github.com/dcampbel24/hnefatafl-copenhagen/compare/v0.4.0...HEAD